Add FieldHandler classes (#12)
Modifies the way fields are created, in order to support using location during validation. This changes handling of fields to be created using a class so that it is possible to have access to common globals in each child field class.

Fields are now managed through a Handler style interface, similar to how views are invoked. This means each fields logic now lives in a handler class which is decided through a map. Each handler defines a get_field method which returns the wtform field to be used. The StringHandler class is the simplest and the FieldHandler is the base that others inherit from. The most complex is the DateHandler which now includes all the logic for date validators within it. Additionally, the date field has been reworked to calculate "dependent" values prior to the field creation. There is further work that could be done in order to define dependencies in a uniform manner throughout runner in order to move the dependency logic completely outside of handlers.