Add clear button functionality for voluntary Radios (#31)
A hidden submit button has been added above the main 
viewable form so that the default action on the 'Enter' 
key is to submit the page rather than clear the Radio 
answers. As part of this work the usage of the 
"save_continue" button action has been removed as the 
default action in the code is to "save and continue".