Use newrelic agent (#40)
This modifies runner to use the preferred newrelic agent on startup. Additionally this prevents library wrapping warnings within newrelic interface. A new make command 'gunicorn' has been added which can start the new relic agent based on appropriate environment vars having been set. A new gunicorn config.py has been added in order to keep settings together as it was impossible to quote vars in a way codacy was happy with. At the same time I've also stripped the server header.

It drops the version of grpcio to the last version, due to the following issue: https://discuss.newrelic.com/t/instrumentation-error-for-python-grpc-with-spanner/91912

Greenlet has also been removed from our dependencies as it does not appear to be used and newrelic will fail to instrument it due to being detected in the environment and never launched

Currently there is a monkeypatching warning shown on startup by gevent but newrelic have indicated it shouldn't be an issue. https://discuss.newrelic.com/t/seeing-monkeypatching-error-while-running-newrelic-admin-with-python-3-7-gevent-and-gunicorn/73893. It appears that gevent is showing this warning in many cases where it need not: gevent/gevent#1239