Add plural rendering (#49)
Adds the ability to use plural variations of a string using babel to look up an appropriate key based on a value. Based on the proposal found [here](https://github.com/ONSdigital/eq-schema-validator/blob/v3/doc/decisions/0003-plural-forms-in-schema-text.md). Added the ability to use the length of a list as a placeholder.