Maintain name duplication state (#277)
Implements capture of household name duplicates. Adds a new key "same_name_items" to the ListStore which holds list_item_ids for all those list items that have a duplicate. Most of the work is done in questionnaire_store_updater.update_same_name_items which updates the duplicates following an update to a list.