Fix 500 error on POST to /questionnaire/relationships (#461)
Sending a POST request to `/questionnaire/relationships` was failing
because it was falling into the non-relationship block handling code.
Although we wouldn't expect this behaviour, it has been observed
during testing, and should be handled properly. To fix this `POST`
has been enabled on the `/questionnaire/relationships` endpoint, which
consequently does the same as the `GET` and redirects to the first
relationship.