Handle HTTP HEAD requests (#573)
Some errors were observed when HTTP HEAD requests were sent to 
questionnaire endpoints. It was identified that this was because a HEAD 
request was following the expected path of a POST request and calling 
form validation. In most instances this didn't result in an error, but in the 
case of an optional Date type answer it did.

The questionnaire changes modify the route logic so that HEAD requests 
follow the same path as a GET request. The session changes modify the 
sign out logic to not sign the user out on a HEAD request.