Fix survey title in header and page title (#622)
To fix the issue the survey title has been added to the cookie on authentication. Although this doesn't align with the decision to not store schema properties in the cookie (https://github.com/ONSdigital/eq-questionnaire-runner/blob/master/doc/architecture/decisions/0010-cookies.md), that would require wider changes to ensure the schema is always in scope when we have a session cookie, and should be done separately.