Use question title as legend (#635)
The design system components now have a legendIsQuestionTitle that removes the need for duplicating the question title (visually hidden) into the answer legend. The rules for when this apply are documented at ons-design-system.netlify.app/patterns/question.

This change implements the logic for everything except Checkbox answers. Checkbox answers don't allow for defining the legend at the schema level - we use the label schema property to populate the checkboxesLabel in the design system macro. This is being looked at by the design system team, and will likely require a new schema property to allow both a label and some kind of instruction (e.g. "Select all that apply").